# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /presets:
    get:
      operationId: listPresets
      tags: [presets]
      summary: List all generation presets
      responses:
        "200":
          description: List of presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  presets:
                    type: array
                    items:
                      $ref: "#/components/schemas/GenerationPreset"
                required:
                  - presets
    post:
      operationId: createPreset
      tags: [presets]
      summary: Create a generation preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePresetRequest"
      responses:
        "201":
          description: Preset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationPreset"
        "409":
          description: Preset name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /presets/{id}:
    get:
      operationId: getPreset
      tags: [presets]
      summary: Get a preset by ID or name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationPreset"
        "404":
          description: Preset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePreset
      tags: [presets]
      summary: Delete a preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Preset deleted
        "404":
          description: Preset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GenerationPreset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        resolution:
          type: string
        fps:
          type: integer
        style:
          type: string
        description:
          type: ["string", "null"]
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - resolution
        - fps
        - style
        - isDefault
        - createdAt
        - updatedAt
    CreatePresetRequest:
      type: object
      properties:
        name:
          type: string
        resolution:
          type: string
          default: "1080p"
        fps:
          type: integer
          default: 24
        style:
          type: string
          default: "cinematic"
        description:
          type: string
        isDefault:
          type: boolean
          default: false
      required:
        - name
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error