# lib/api-zod/src/generated/types/createPresetRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePresetRequest {
  name: string;
  resolution?: string;
  fps?: number;
  style?: string;
  description?: string;
  isDefault?: boolean;
}