# lib/api-zod/src/generated/types/generationPreset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationPreset {
  id: string;
  name: string;
  resolution: string;
  fps: number;
  style: string;
  /** @nullable */
  description?: string | null;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}